--- src/components/Button/index.tsx
@@ -1,2 +1,0 @@
-export * from "./Colored";
-export * from "./Outline";
