--- src/components/Input/index.tsx
@@ -1,9 +1,0 @@
-import * as Basic from "./Basic";
-import * as Checkbox from "./Checkbox";
-import * as Select from "./Select";
-
-export default {
-  ...Basic,
-  ...Checkbox,
-  ...Select,
-};
