+++ src/main.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { AuthContextProvider } from "./contexts/AuthContext";
 
 import Home from "./pages/Home";
 import Login from "./pages/Login";
@@ -8,7 +9,6 @@ import KakaoLogin, { kakaoLoader } from "./pages/KakaoLogin";
 import SignUp from "./pages/SignUp";
 
 import "./styles/index.scss";
-import { AuthContextProvider } from "./contexts/AuthContext";
 
 const router = createBrowserRouter([
   {
