+++ tailwind.config.cjs
@@ -5,12 +5,15 @@ module.exports = {
     extend: {
       colors: {
         primary: "#F4B183",
+        "primary-peach": "#FBE5D6",
         "grayscale-100": "#000000",
         "grayscale-80": "#696969",
+        "grayscale-60": "#808080",
         "grayscale-20": "#c0c0c0",
         "system-alert": "#ff0000",
         "system-yellow": "#FFE699",
         "system-dimyellow": "#F8F5D5",
+        "kakao-container": "#FEE500",
       },
     },
   },
